refactor: change qos naming to monitoring
BREAKING CHANGE: Updated all package names and configuration from `pillarbox.qos` to
`pillarbox.monitoring`. This includes changes to environment variables, which now use the
prefix `PILLARBOX_MONITORING_*` instead of `PILLARBOX_QOS_*`. Ensure that all relevant
configurations are updated to reflect this change when deploying or running the service.
