+++ src/main/resources/application.yml
@@ -2,8 +2,8 @@ spring:
   threads.virtual.enabled: true
   application.name: pillarbox-monitoring-transfer
 
-pillarbox.qos:
-  sse.uri: "http://localhost:8080/events"
+pillarbox.monitoring:
+  dispatch.uri: "http://localhost:8080/events"
   opensearch.uri: "http://localhost:9200"
 
 server:
@@ -17,4 +17,4 @@ management:
   health.benchmark.enabled: true
 
 logging.level:
-  ch.srgssr.pillarbox.qos: INFO
+  ch.srgssr.pillarbox.monitoring: INFO
