+++ src/test/kotlin/ch/srgssr/pillarbox/monitoring/PillarboxQosDataTransferApplicationTests.kt
@@ -1,0 +1,12 @@
+package ch.srgssr.pillarbox.monitoring
+
+import org.junit.jupiter.api.Test
+import org.springframework.boot.test.context.SpringBootTest
+
+@SpringBootTest
+class PillarboxQosDataTransferApplicationTests {
+  @Test
+  @Suppress("EmptyFunctionBlock")
+  fun contextLoads() {
+  }
+}
