--- src/test/kotlin/ch/srgssr/pillarbox/qos/PillarboxQosDataTransferApplicationTests.kt
@@ -1,12 +1,0 @@
-package ch.srgssr.pillarbox.qos
-
-import org.junit.jupiter.api.Test
-import org.springframework.boot.test.context.SpringBootTest
-
-@SpringBootTest
-class PillarboxQosDataTransferApplicationTests {
-    @Test
-    @Suppress("EmptyFunctionBlock")
-    fun contextLoads() {
-    }
-}
